requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.36.1
botocore==1.36.1
cachetools==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dataclasses-json==0.6.7
dill==0.3.9
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
free_proxy==1.1.3
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google_search_results==2.4.2
googlesearch-python==1.2.5
greenlet==3.1.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.19
langchain-aws==0.2.11
langchain-community==0.3.18
langchain-core==0.3.41
langchain-mistralai==0.2.4
langchain-ollama==0.2.2
langchain-openai==0.3.7
langchain-postgres==0.0.13
langchain-text-splitters==0.3.6
langgraph==0.3.2
langgraph-checkpoint==2.0.16
langgraph-prebuilt==0.1.1
langgraph-sdk==0.1.53
langsmith==0.3.11
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
mdurl==0.1.2
minify_html==0.15.0
mpire==2.10.2
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
narwhals==1.22.0
numpy==2.2.3
ollama==0.4.6
openai==1.59.7
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pgvector==0.3.6
pillow==11.1.0
playwright==1.49.1
promptimal==2.0.0
propcache==0.2.1
protobuf==5.29.3
psycopg==3.2.5
psycopg-pool==3.2.6
psycopg2 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_036gyhqc5v/croot/psycopg2_1731687658365/work
psycopg2-binary==2.9.10
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.0.0
Pygments==2.19.1
pypdf==5.3.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.11.1
scrapegraphai==1.36.0
semchunk==3.0.1
serpapi==0.1.5
simpleeval==1.0.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
streamlit==1.41.1
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
undetected-playwright==0.3.0
urllib3==2.3.0
urwid==2.6.16
wcwidth==0.2.13
yarl==1.18.3
zstandard==0.23.0